+++ src/MF_Segment/MFSegments.h
@@ -14,7 +14,7 @@ class MFSegments
 public:
     MFSegments();
     void display(uint8_t module, char *string, uint8_t points, uint8_t mask, bool convertPoints = false);
-    void attach(uint8_t type, int dataPin, int csPin, int clkPin, uint8_t moduleCount, uint8_t brightness);
+    void attach(uint8_t type, uint8_t dataPin, uint8_t csPin, uint8_t clkPin, uint8_t moduleCount, uint8_t brightness);
     void detach();
     void test();
     void powerSavingMode(bool state);
